Move document spell check handler
